[#3]Refactor: 라우터 설정을 router.jsx로 분리
